Inline headers from base, comm, random, and sim
