+++ src/sim/SimTrackView.hh
@@ -57,6 +57,74 @@ class SimTrackView
 };
 
 //---------------------------------------------------------------------------//
-} // namespace celeritas
+// INLINE DEFINITIONS
+//---------------------------------------------------------------------------//
+/*!
+ * Construct from persistent and local data.
+ */
+CELER_FUNCTION
+SimTrackView::SimTrackView(const SimStateRef& states, ThreadId thread)
+    : states_(states), thread_(thread)
+{
+    CELER_EXPECT(thread < states_.size());
+}
 
-#include "SimTrackView.i.hh"
+//---------------------------------------------------------------------------//
+/*!
+ * \brief Initialize the particle.
+ */
+CELER_FUNCTION SimTrackView& SimTrackView::operator=(const Initializer_t& other)
+{
+    states_.state[thread_] = other;
+    return *this;
+}
+
+//---------------------------------------------------------------------------//
+// DYNAMIC PROPERTIES
+//---------------------------------------------------------------------------//
+/*!
+ * Unique track identifier.
+ */
+CELER_FUNCTION TrackId SimTrackView::track_id() const
+{
+    return states_.state[thread_].track_id;
+}
+
+//---------------------------------------------------------------------------//
+/*!
+ * Track ID of parent.
+ */
+CELER_FUNCTION TrackId SimTrackView::parent_id() const
+{
+    return states_.state[thread_].parent_id;
+}
+
+//---------------------------------------------------------------------------//
+/*!
+ * Event ID.
+ */
+CELER_FUNCTION EventId SimTrackView::event_id() const
+{
+    return states_.state[thread_].event_id;
+}
+
+//---------------------------------------------------------------------------//
+/*!
+ * Whether the track is alive.
+ */
+CELER_FUNCTION bool SimTrackView::alive() const
+{
+    return states_.state[thread_].alive;
+}
+
+//---------------------------------------------------------------------------//
+/*!
+ * Set whether the track is alive.
+ */
+CELER_FUNCTION void SimTrackView::alive(bool is_alive)
+{
+    states_.state[thread_].alive = is_alive;
+}
+
+//---------------------------------------------------------------------------//
+} // namespace celeritas
